@@ -4,13 +4,15 @@
   "private": true,
   "dependencies": {
     "@ant-design/icons": "^5.3.7",
+    "@testing-library/jest-dom": "^6.4.6",
     "antd": "^5.18.1",
-    "axios": "^0.25.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.3"
+    "axios": "1.7.2",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-hot-toast": "^2.4.1",
+    "react-router-dom": "6.23.1",
+    "react-scripts": "^5.0.1",
+    "styled-components": "6.1.11"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -24,6 +26,11 @@
       "react-app/jest"
     ]
   },
+  "jest": {
+    "transformIgnorePatterns": [
+      "node_modules/(?!axios)/"
+    ]
+  },
   "browserslist": {
     "production": [
       ">0.2%",
@@ -35,5 +42,16 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@babel/core": "^7.24.7",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@babel/preset-env": "^7.24.7",
+    "@babel/preset-react": "^7.24.7",
+    "@testing-library/dom": "^10.2.0",
+    "@testing-library/react": "^16.0.0",
+    "babel-jest": "^29.7.0",
+    "jest": "^28.0.0",
+    "jest-mock-axios": "^4.7.3"
   }
 }
